requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
python-dotenv==1.0.1
Werkzeug==3.1.3
aiohttp
aiosignal
annotated-types
anyio
appnope
asttokens
attrs
beautifulsoup4
certifi
charset-normalizer
click
comm
dataclasses-json
debugpy
decorator
distro
docx2txt
duckduckgo_search
executing
frozenlist
greenlet
h11
httpcore
httpx
idna
ipykernel
ipython
jedi
jsonpatch
jsonpointer
jupyter_client
jupyter_core
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langchainhub
langsmith
marshmallow
matplotlib-inline
multidict
mypy-extensions
nest-asyncio
numpy
openai
orjson
packaging
pandas
parso
pexpect
pinecone-client
pinecone-plugin-interface
platformdirs
prompt_toolkit
psutil
ptyprocess
pure-eval
pydantic
pydantic_core
Pygments
pypdf
pyreqwest_impersonate
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
regex
requests
six
sniffio
soupsieve
SQLAlchemy
stack-data
tenacity
tiktoken
tk
tornado
tqdm
traitlets
types-requests
typing-inspect
typing_extensions
tzdata
urllib3
wcwidth
wikipedia
yarl
